requirements.txt
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.3.0
streamlit==1.28.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
google-generativeai==0.3.1
python-dotenv==1.0.0
surprise==0.1
plotly==5.18.0
nltk==3.8.1
tqdm==4.66.1
joblib==1.3.2
